refactor : component 리팩토링
